Split InfiltratorGameModule into multiple files
