+++ DiscordBot/Infiltrator/Player.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Threading.Tasks;
+using Discord;
+using Discord.Commands;
+
+namespace DiscordBot.Infiltrator
+{
+
+    public class Player
+    {
+
+    }
+}
